Huge autoload refactor
