+++ src/Entity/DfCmsBuild.php
@@ -13,7 +13,7 @@
  * @license   GPLv3
  */
 
-namespace PrestaShop\Module\Doofinder\Src\Entity;
+namespace PrestaShop\Module\Doofinder\Entity;
 
 if (!defined('_PS_VERSION_')) {
     exit;
