+++ src/Entity/DoofinderResults.php
@@ -12,7 +12,7 @@
  * - getResults: get an array with the results
  */
 
-namespace PrestaShop\Module\Doofinder\Src\Entity;
+namespace PrestaShop\Module\Doofinder\Entity;
 
 if (!defined('_PS_VERSION_')) {
     exit;
